refactor: extract games_stream.py from games.py
Move stream_turn SSE endpoint + event_stream generator + _emit helper
to games_stream.py (262 lines).

games.py: 596 → 351 lines (-245, -41.1%)
Cumulative: 1610 → 351 (games.py) + 79 + 374 + 371 + 317 + 262 (extracted)

All 6 modules under 400 lines. 11 routes intact, 75/75 tests pass.
